Made a concept system to create a tile node system for navigation
This needs to be tweaked for performance!
Current idea: Make an object called nodes. This stores "Passable" as a boolean and it stores it's coordinates. The dictionary would then be changed to <float, List<Node>>. Most of the same concepts would apply. Maybe we can make an invisible cube and place it where each node is supposed to be? The idea would be to make the collider for the nodes as a trigger. When something touches them, we mark them as impassable when they are missing the tag "Passable". Just an idea.