Node Clusters!
Finished some tweaks to nodes so they can start being grouped into clusters. Currently, there's no method to make a cluster regardless of it being contested or not. We're going to have to find a way to handle creating general clusters.

The neat part is this can be used to create walls. It may need a tweak, but the theory behind it is basically the same
Don't want the agent to walk in to the wall? Give it a distance offset, it'll take distance offset + collider size - position for the start. Loop through all nodes from there to offset + collider size + position.
If it's contested by the wall, then add it. And the wall can be identified in two ways. One with a simple script that probably will inherit from a class "TerrainObject" which would be inherited by TerrainNavigator as well (Or the agent). You can put the distance offset to this and it'll work with the agents as well.
TerrainObject will track that and the cluster of nodes that belong to it. Probably a few other fields as well.
The other method of this would be create a layer that's "TerrainObstacle"
When generating the terrain, you check all nodes that have a TerrainObstacle in it, then add that to a cluster of unwalkable nodes.
Both may be a good option as well. But I need to test more stuff